database/enquiry --> Modified Enquiry Table

+++ app/routes/enquiries.py
@@ -3,7 +3,7 @@ from sqlalchemy import select
 from sqlalchemy.orm import Session
 
 from app.database import get_db
-from app.dependencies import get_current_user
+from app.dependencies import (get_current_user, get_optional_current_user)
 from app.models import Enquiry, User, UserWishlist
 from app.schemas import (
     EnquiryCreate,
@@ -56,36 +56,56 @@ def get_enquiries(
 )
 def create_enquiry(
     data: EnquiryCreate,
-    current_user: User = Depends(get_current_user),
+    current_user: User | None = Depends(get_optional_current_user),
     database: Session = Depends(get_db),
 ):
     enquiry = Enquiry(
-        user_id=current_user.id,
+        user_id=(
+            current_user.id
+            if current_user is not None
+            else None
+        ),
         product_id=data.product_id,
         product_slug=data.product_slug,
         product_name=data.product_name,
         customer_name=data.customer_name,
         email=data.email,
         phone=data.phone,
+        city=data.city,
+        selected_addons=data.selected_addons,
         message=data.message,
         status="submitted",
     )
 
     database.add(enquiry)
 
-    wishlist_statement = select(
-        UserWishlist
-    ).where(
-        UserWishlist.user_id == current_user.id,
-        UserWishlist.product_id == data.product_id,
-    )
+    # wishlist_statement = select(
+    #     UserWishlist
+    # ).where(
+    #     UserWishlist.user_id == current_user.id,
+    #     UserWishlist.product_id == data.product_id,
+    # )
 
-    wishlist_item = database.scalar(
-        wishlist_statement
-    )
+    # wishlist_item = database.scalar(
+    #     wishlist_statement
+    # )
 
-    if wishlist_item is not None:
-        database.delete(wishlist_item)
+    # if wishlist_item is not None:
+    #     database.delete(wishlist_item)
+    if current_user is not None:
+        wishlist_statement = select(
+            UserWishlist
+        ).where(
+            UserWishlist.user_id == current_user.id,
+            UserWishlist.product_id == data.product_id,
+        )
+
+        wishlist_item = database.scalar(
+            wishlist_statement
+        )
+
+        if wishlist_item is not None:
+            database.delete(wishlist_item)
 
     database.commit()
     database.refresh(enquiry)
